actually commit everything
